QPID-1055: use int64 for encoding python both python int and longs; this ensures consistent behavior on both 64 bit and non 64 bit systems
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@655951 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/python/qpid/spec010.py
@@ -447,7 +447,7 @@ class Spec(Node):
 
   ENCODINGS = {
     basestring: "vbin16",
-    int: "int32",
+    int: "int64",
     long: "int64",
     None.__class__: "void",
     list: "list",
